+++ tests/test_widgets.py
@@ -251,6 +251,7 @@ class TestFrame6(Frame):
 class TestWidgets(unittest.TestCase):
     def setUp(self):
         self.maxDiff = None
+        self.epoch_date = str(date.fromtimestamp(0))
 
     def assert_canvas_equals(self, canvas, expected):
         """
@@ -1964,9 +1965,9 @@ class TestWidgets(unittest.TestCase):
         self.assert_canvas_equals(
             canvas,
             "/                     Size Last modified\n" +
-            "|-- A Directory          0    1970-01-01\n" +
-            "|-- A File               0    1970-01-01\n" +
-            "|-- A Lnk                0    1970-01-01\n" +
+            "|-- A Directory          0    %s\n" % self.epoch_date +
+            "|-- A File               0    %s\n" % self.epoch_date +
+            "|-- A Lnk                0    %s\n" % self.epoch_date +
             "                                        \n" +
             "                                        \n" +
             "                                        \n" +
@@ -2054,11 +2055,11 @@ class TestWidgets(unittest.TestCase):
         self.assert_canvas_equals(
             canvas,
             "/                     Size Last modified\n" +
-            "|-+ A Directory         9K    1970-01-01\n" +
-            "|-+ Lnk Directo...      9K    1970-01-01\n" +
-            "|-- A File              9K    1970-01-01\n" +
-            "|-- A Lnk -> A Tgt      9K    1970-01-01\n" +
-            "|-- oööÖ.txt            9K    1970-01-01\n" +
+            "|-+ A Directory         9K    %s\n" % self.epoch_date +
+            "|-+ Lnk Directo...      9K    %s\n" % self.epoch_date +
+            "|-- A File              9K    %s\n" % self.epoch_date +
+            "|-- A Lnk -> A Tgt      9K    %s\n" % self.epoch_date +
+            "|-- oööÖ.txt            9K    %s\n" % self.epoch_date +
             "                                        \n" +
             "                                        \n" +
             "                                        \n" +
@@ -2087,11 +2088,11 @@ class TestWidgets(unittest.TestCase):
             canvas,
             "/A Directory          Size Last modified\n" +
             "|-+ ..                                  \n" +
-            "|-+ A Directory         9K    1970-01-01\n" +
-            "|-+ Lnk Directo...      9K    1970-01-01\n" +
-            "|-- A File              9K    1970-01-01\n" +
-            "|-- A Lnk -> A Tgt      9K    1970-01-01\n" +
-            "|-- oööÖ.txt            9K    1970-01-01\n" +
+            "|-+ A Directory         9K    %s\n" % self.epoch_date +
+            "|-+ Lnk Directo...      9K    %s\n" % self.epoch_date +
+            "|-- A File              9K    %s\n" % self.epoch_date +
+            "|-- A Lnk -> A Tgt      9K    %s\n" % self.epoch_date +
+            "|-- oööÖ.txt            9K    %s\n" % self.epoch_date +
             "                                        \n" +
             "                                        \n" +
             "                                        \n")
@@ -2147,8 +2148,8 @@ class TestWidgets(unittest.TestCase):
         self.assert_canvas_equals(
             canvas,
             "/                     Size Last modified\n" +
-            "|-+ A Directory         9K    1970-01-01\n" +
-            "|-- hello.bmp           9K    1970-01-01\n" +
+            "|-+ A Directory         9K    %s\n" % self.epoch_date +
+            "|-- hello.bmp           9K    %s\n" % self.epoch_date +
             "                                        \n" +
             "                                        \n" +
             "                                        \n" +
@@ -3360,6 +3361,109 @@ class TestWidgets(unittest.TestCase):
         # by confirming the label's value
         l = form.find_widget("tf6_lbl")
         self.assertEqual(l.text, "TstFrm6Lbl")
+
+    def test_textbox_hide_cursor(self):
+        """
+        Check hide cursor works as expected.
+        """
+        # Now set up the Frame ready for testing
+        screen = MagicMock(spec=Screen, colours=8, unicode_aware=False)
+        scene = Scene([], duration=-1)
+        canvas = Canvas(screen, 10, 40, 0, 0)
+        form = Frame(canvas, canvas.height, canvas.width)
+        layout = Layout([100])
+        form.add_layout(layout)
+        textbox = TextBox(4, label="TB")
+        layout.add_widget(textbox)
+        form.fix()
+        form.register_scene(scene)
+        scene.add_effect(form)
+        scene.reset()
+
+        # Check we default to showing the cursor.
+        self.assertFalse(textbox.hide_cursor)
+
+        # Check that the cursor is hidden when set off.
+        for effect in scene.effects:
+            effect.update(0)
+        self.assertEqual(canvas.get_from(4, 1),
+                         (32, Screen.COLOUR_WHITE, Screen.A_REVERSE, Screen.COLOUR_CYAN))
+        textbox.hide_cursor = True
+        for effect in scene.effects:
+            effect.update(1)
+        self.assertEqual(canvas.get_from(4, 1),
+                         (32, Screen.COLOUR_WHITE, Screen.A_BOLD, Screen.COLOUR_CYAN))
+
+    def test_textbox_autoscroll(self):
+        """
+        Check autoscroll works as expected.
+        """
+        # Now set up the Frame ready for testing
+        screen = MagicMock(spec=Screen, colours=8, unicode_aware=False)
+        scene = Scene([], duration=-1)
+        canvas = Canvas(screen, 10, 40, 0, 0)
+        form = Frame(canvas, canvas.height, canvas.width)
+        layout = Layout([100])
+        form.add_layout(layout)
+        textbox = TextBox(4, label="TB")
+        textbox.value = ["Hello", "World", "Full", "Box"]
+        layout.add_widget(textbox)
+        form.fix()
+        form.register_scene(scene)
+        scene.add_effect(form)
+        scene.reset()
+
+        # Check that the frame is rendered correctly.
+        for effect in scene.effects:
+            effect.update(0)
+        self.assert_canvas_equals(
+            canvas,
+            "+--------------------------------------+\n" +
+            "|TB Hello                              |\n" +
+            "|   World                              O\n" +
+            "|   Full                               |\n" +
+            "|   Box                                |\n" +
+            "|                                      |\n" +
+            "|                                      |\n" +
+            "|                                      |\n" +
+            "|                                      |\n" +
+            "+--------------------------------------+\n")
+
+        # Check that switching off autoscroll will not move line
+        textbox.auto_scroll = False
+        textbox.value = ["Hello", "World", "Full", "Box", "New"]
+        for effect in scene.effects:
+            effect.update(0)
+        self.assert_canvas_equals(
+            canvas,
+            "+--------------------------------------+\n" +
+            "|TB Hello                              |\n" +
+            "|   World                              O\n" +
+            "|   Full                               |\n" +
+            "|   Box                                |\n" +
+            "|                                      |\n" +
+            "|                                      |\n" +
+            "|                                      |\n" +
+            "|                                      |\n" +
+            "+--------------------------------------+\n")
+
+        # Check that turning it back on scrolls to bottom.
+        textbox.auto_scroll = True
+        textbox.value = ["Hello", "World", "Full", "Box", "New"]
+        for effect in scene.effects:
+            effect.update(0)
+        self.assert_canvas_equals(
+            canvas,
+            "+--------------------------------------+\n" +
+            "|TB World                              |\n" +
+            "|   Full                               O\n" +
+            "|   Box                                |\n" +
+            "|   New                                |\n" +
+            "|                                      |\n" +
+            "|                                      |\n" +
+            "|                                      |\n" +
+            "|                                      |\n" +
+            "+--------------------------------------+\n")
 
 
 if __name__ == '__main__':