+++ CHANGES.rst
@@ -8,6 +8,7 @@ LATEST
 - Added fit parameter to DropdownList.
 - Added support for default colours to AnsiTerminalParser
 - BREAKING VISUAL CHANGE: Frame now supports scroll bars without borders, to have no border and no scroll bar you now need Frame(has_border=False, can_scroll=False)
+- Added TextBox.hide_cursor and TextBox.auto_scroll properties
 
 1.13.0
 ------